# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Services"
	},
	"paths": {
		"/users/register": {
			"post": {
				"summary": "Singup of the user",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User",
							"type": "object",
							"properties": {
								"username": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"type": "object",
							"properties": {
								"status": {
									"type": "string"
								}, "message": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Failed operation",
						"schema": {
							"$ref": "#/definitions/failure"
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"summary": "Authenticates the users",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"username": {
									"type": "string"
								}, "password": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"type": "object",
							"properties": {
								"status": {
									"type": "string"
								}, "token": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Failed operation",
						"schema": {
							"$ref": "#/definitions/failure"
						}
					}
				}
			}
		},
		"/products": {
			"post": {
				"summary": "Endpoint to upload products. This endpoint is restricted to admin access.",
				"consumes": [
					"multipart/form-data"
				],
				"produces": [
					"application/json"
				],
				"parameters": [{
					"in": "formData",
					"name": "products",
					"required": true,
					"type": "file",
					"x-mimetype": "application/json"
				}, {
					"in": "header",
					"name": "x-access-token",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"type": "array",
							"items": {
								"type": "object",
								"properties": {
									"status": {
										"type": "string"
									}, "token": {
										"type": "string"
									}
								}
							}
						}
					}, 
					"400": {
						"description": "Failed operation",
						"schema": {
							"$ref": "#/definitions/failure"
						}
					}
				}
			}
		},
		"/products/review": {
			"post": {
				"summary": "Client reviews an existing product in the system. This endpoint is restricted to client access.",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"in": "body",
					"name": "body",
					"required": true,
					"schema": {
						"type": "object",
						"properties": {
							"barcode": {
								"type": "number"
							}, "review": {
								"type": "string"
							}
						}
					}
				}, {
					"in": "header",
					"name": "x-access-token",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/failure"
						}
					}, 
					"400": {
						"description": "Failed operation",
						"schema": {
							"$ref": "#/definitions/failure"
						}
					}
				}
			}
		},
		"/products/search": {
			"post": {
				"summary": "Search for the desired product by name in the system. This endpoint is restricted to client access.",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"in": "query",
					"name": "searchText",
					"required": true,
					"type": "string"
				}, {
					"in": "query",
					"name": "page",
					"required": true,
					"type": "number"
				}, {
					"in": "header",
					"name": "x-access-token",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"type": "object",
							"properties": {
								"status": {
									"type": "string"
								},
								"message": {
									"type": "string"
								}, 
								"totalCount": {
									"type": "number"
								}, 
								"products": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"_id": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"barcode": {
												"type": "number"
											},
											"brand": {
												"type": "string"
											},
											"description": {
												"type": "string"
											},
											"price": {
												"type": "string"
											},
											"available": {
												"type": "boolean"
											}, 
											"reviews": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"name": {
															"type": "string"
														},
														"review": {
															"type": "string"
														}
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Failed operation",
						"schema": {
							"$ref": "#/definitions/failure"
						}
					}
				}
			}
		}
	},
	"definitions": {
		"User": {
			"type": "object",
			"properties": {
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}, 
				"name": {
					"type": "string"
				},
				"role": {
					"type": "string",
					"enum": [ "client", "admin" ]
				},
				"email": {
					"type": "string"
				},
				"mobileNo": {
					"type": "number"
				},
				"gender": {
					"type": "string",
					"enum": ["M", "F"]
				}
			}
		},
		"Products": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"barcode": {
					"type": "number"
				},
				"brand": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"price": {
					"type": "number"
				},
			"available": {
				"type": "boolean"
			}
		}
	},
	"reviews": {
		"type": "object",
		"properties":{
			"id": {
				"type": "integer",
				"format": "int64"
			},
			"userId": {
				"$ref": "#/definitions/User"
			},
			"review": {
				"type": "string"
			}
		}
	},
	"failure": {
		"type": "object",
		"properties":{
			"status": {
				"type": "string"
			},
			"message": {
				"type": "string"
			}
		}
	}
}

}